PlotData is separated to a new file. add the new func convertArrayToPlotData
